Started changing the question system so that you can not just check off system questions but you can really add and edit all details of questions, for all article types and for members. Have this working for story questions, need to do it for all other types as well.  It is not that hard to change this now, and the way I had it was too simple to even begin to play with it fully. This is MUCH more interesting.
Also moved community questions and rules to snippets because they are not going to work for a while and it's better to have only working code in the code files. Snippets is a repository of stuff I abandoned, just in case I figured out something useful in that code and need it again.