Retelling and reminding (story to story) now working. Related and included links can be displayed but cannot yet be created :)
Also, changed how members join. Instead of adding full members, managers/owners add email addresses as pending members. The first time a pending member visits the community, the pending member object goes away and a full member object is created. The only bad thing about this is that if the person changes the email on their Google account in the interim they will not be able to log in. However that is pretty unlikely and there is not really any other good way to do it. The method I got from the web to look up Google accounts from emails didn't work on the development server, and might have online, but I was nervous about it anyway. This is cleaner.

Next up is member pages and histories.