Have been wrestling with getting the correct time (on entries) to show for the user's time zone. Finally have this working in most places, using pytz, but still need to set up asking the user for what time zone to show (and the community for what default to set). Am not checking in pytz but need to write some instructions for setting this up (now two extra packages are required, gaeutilities and pytz).
Had a lot of trouble with a bug in the GAE launcher where if you load a Python zipped "egg" it chokes. Stopped doing that and put the pytz code right in the directory with things. Pytz has 500 or so files (one per time zone) and so people say it's a problem with the 1000 file GAE limit. I tried a method where people zipped the time zone info files, but I thought it slowed things down, and I don't have that many files. If for some reason the number of files becomes an issue later I can use the zipped version instead.

Also finally figured out how to make custom django template filters, and that solves the horrid hard-coding-index-numbers problem, so I can go back and fix all of those places later. Many small things left to do but things are shaping up!