Made "review resources" page for guides. Can look over resources and flag them for removal (same system as curators, but only for resources, and cannot send notifications, because resources can only be entered by guides and managers/owners, so they are probably not going to be silly).
Also, realized that it is actually pretty easy to make it possible to re-enter ANY published article, so put in a community option. Now the community owners/managers can decide whether the person who created the article (of any type they specify) can re-enter it even after it is published and make changes. Probably people will only turn this on for non-stories, or that is the recommendation. I am NOT going to implement this for annotations, answers or links, because it could just go on and on in complexity. Still, for systemic things like collages and resources, re-entry could be very useful.