Working on the paging issue. First, for the four grids (home, entry, member, character), have standardized the display somewhat so that you see the same components on each screen. Improved the look of the selection boxes to save space. Also, added more options to reduce the number of things you are looking at for entry, member and character.
Most importantly there is a max number of items to be displayed per page, which can be set by the admin but should always be (way) less than the real limit of 1000. The default is 200 items (mainly links) to be shown per page. If the number of items is more than the max, the user gets a warning note that only so many OF THE MOST RECENT items have been shown. The user can show older items by making more selections to trim the number of items down below the max. So, if they are trying to look at all of a member's contributions and there are more than the max of (say) 200, it will tell them that, and they can see older items by not looking at tags or nudges or whatever. This is an attempt to make limited queries work with the grid display WITHOUT putting up the dreaded "stories 125-150 of 4652" message which I hate. We'll see how it plays out in practice. (For other non-grid lists simple paging is fine.)