Three main changes. First, added a "simple search" function where it just looks everywhere for words entered. Nothing is saved, this is for on the fly searching, not like filters which are more complex.
Second, added a thing where it says "Your changes have been saved" in every instance where the user goes BACK to a page after making a change. Note, it doesn't actually know whether they changed anything, it just knows that they clicked a button that landed it in a POST method. Checking whether any data actually changed would be much harder. 

Third, added a utility to check the correspondence between the help.csv file and the templates, because they can get out of alignment if you are messing with either.