Added code that checks which application ID is being used before it sets things. This is only used right now to dynamically the French version which is being tested. The reason to add this dynamic app_id-dependent setting is that otherwise I would have to change the code (SITE_LANGUAGE="francais"), upload, change the code again (SITE_LANGUAGE="english"), upload, etc, for each minor-version change. If I did things out of sequence the English sites could suddenly turn French! This will help anybody deploying multiple language versions at the same time.
Also, changed it so that there is a "fall-back" language directory (usually "english") which is used if any files are missing from the specified language directory. This is to show partially completed translations so that the person translating can see what they have done so far working.

Did NOT increment the uploaded version number, but will bundle this smallish change into the next minor version which will go up soon with some other changes. 

This check-in includes the first file (default_resources.txt) of the French translation. THANKS STEPHANE!!